feat: added build instructions for SGPT from source code and docker
This commit adds instructions on how to build SGPT from the source code using Go and how to build a Docker image for SGPT. The `README.md` file has been updated with the new instructions. The `bin/docker-build.sh` script has been modified to use the `latest` tag for the Docker image.