fix: Save Chat Sessions by Mounting User Home Dir in Docker
This commit fixes an issue when using the SGPT Docker image where chat sessions were not being saved due to a missing directory. To address this, I added a `VOLUME /home/nonroot` line to the Dockerfile to ensure that the `nonroot` user's home directory is properly mounted as a volume. Additionally, I updated the `internal/chat` package to create the necessary directory structure if it does not already exist. With these changes, users can now mount a local folder to the `nonroot` user's home directory and persist all active chat sessions across different container instances.