chore: Update development instructions for SGPT
In this commit, we updated the development instructions for SGPT. Specifically, we removed redundant information about building from source code and changed the header for building with Go to a sub-section. Additionally, we changed the header for building a Docker image to be on the same level as building with Go. These changes improve the clarity and organization of the SGPT development instructions. No functionality was changed in this commit.