feat:replace ffcli with cobra (#40)
* chore: refactor package structure

* feat: implement cli with cobra lib

* feat: introduce sgpt client

* chore: remove openai lib in cli package

* chore: rename getprompt to getinput

* chore: move shell specific commands to shell package

* chore: refactor image saving action

a prefix can be set via a flag for the filename. the suffix is always randomly generated

* chore: refactor folder structure

* chore: refactor move chat file func to files

* chore: refactor, rename files to filesystem package, modfifier to modifiers, rename vars

* chore: remove unused dependencies

* fix(cli): improved command line interface for shell command

The commit changed the behavior of the shell command by adding new options to improve usability. The options include `model`, `max-tokens`, `temperature`, `top-p`, and `chat`, which can be specified using shorthand flags `-m`, `-s`, `-t`, `-p`, and `-c`, respectively. Additionally, the commit fixed a bug that prevented shell commands from being executed when the `execute` flag is set to `true`. The fix also added a fallback behavior of printing the response to stdout if execution fails.