feat: Add additional pre-commit hooks
Added a pre-commit hook to enforce the use of a valid SPDX license identifier in all files. Also made updates to the `.mdl_config.yaml` file to include a copyright notice and a SPDX-License-Identifier. This commit aligns with the guidelines of semantic versioning, as it introduces new features to the project.